chore: split workspace report rendering
