+++ cli/python/base_projects/workspace_reports.py
@@ -14,12 +14,16 @@ from base_projects.workspace_scanner import workspace_manifest_entries
 from base_projects.workspace_manifest import WorkspaceManifest
 from base_projects.workspace_manifest import WorkspaceManifestRepo
 from base_projects.workspace_manifest import read_workspace_manifest
+from base_projects.workspace_report_json import dumps_json  # pylint: disable=unused-import
+from base_projects.workspace_report_json import workspace_check_to_json  # pylint: disable=unused-import
+from base_projects.workspace_report_json import workspace_doctor_to_json  # pylint: disable=unused-import
+from base_projects.workspace_report_json import workspace_status_to_json  # pylint: disable=unused-import
+from base_projects.workspace_report_text import print_workspace_check  # pylint: disable=unused-import
+from base_projects.workspace_report_text import print_workspace_doctor  # pylint: disable=unused-import
+from base_projects.workspace_report_text import print_workspace_status  # pylint: disable=unused-import
 from base_setup.checks import ArtifactCheck
-from base_setup.checks import DIAGNOSTIC_JSON_SCHEMA_VERSION
-from base_setup.checks import check_to_json
 from base_setup.checks import checks_status
 from base_setup.checks import doctor_status
-from base_setup.checks import print_doctor_finding
 from base_setup.engine import manifest_checks
 from base_setup.engine import pre_venv_manifest_checks
 from base_setup.engine import read_default_manifest
@@ -545,281 +549,3 @@ def project_venv_check(manifest: BaseManifest) -> ArtifactCheck:
 
 def workspace_error_count(results: tuple[WorkspaceProjectCheckResult, ...]) -> int:
     return sum(1 for result in results for check in result.checks if doctor_status(check) == "error")
-
-
-def workspace_status_to_json(
-    workspace_root: Path,
-    statuses: tuple[WorkspaceProjectStatus, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> dict[str, Any]:
-    payload: dict[str, Any] = {
-        "workspace": str(workspace_root),
-        "project_count": workspace_project_count(statuses, workspace_manifest),
-        "projects": [workspace_status_item_to_json(status, workspace_manifest) for status in statuses],
-    }
-    add_workspace_manifest_json(payload, statuses, workspace_manifest)
-    return payload
-
-
-def workspace_status_item_to_json(
-    status: WorkspaceProjectStatus,
-    workspace_manifest: WorkspaceManifest | None,
-) -> dict[str, Any]:
-    item: dict[str, Any] = {
-        "name": status.name,
-        "status": status.status,
-        "path": str(status.root),
-        "manifest_path": str(status.manifest_path) if status.manifest_path is not None else None,
-        "venv": status.venv,
-        "manifest": status.manifest,
-        "last_check": last_check_to_json(status.last_check),
-        "issues": list(status.issues),
-    }
-    if workspace_manifest is not None:
-        item.update(workspace_manifest_item_metadata(status))
-    if status.python_runtime is not None:
-        item["python_runtime"] = status.python_runtime.to_json()
-    return item
-
-
-def workspace_check_to_json(
-    workspace_root: Path,
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> dict[str, Any]:
-    return workspace_checks_to_json(workspace_root, results, workspace_manifest=workspace_manifest)
-
-
-def workspace_doctor_to_json(
-    workspace_root: Path,
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> dict[str, Any]:
-    return workspace_checks_to_json(workspace_root, results, workspace_manifest=workspace_manifest)
-
-
-def workspace_checks_to_json(
-    workspace_root: Path,
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> dict[str, Any]:
-    payload: dict[str, Any] = {
-        "schema_version": DIAGNOSTIC_JSON_SCHEMA_VERSION,
-        "workspace": str(workspace_root),
-        "status": checks_status(tuple(check for result in results for check in result.checks)),
-        "project_count": workspace_project_count(results, workspace_manifest),
-        "projects": [workspace_check_result_to_json(result, workspace_manifest) for result in results],
-    }
-    add_workspace_manifest_json(payload, results, workspace_manifest)
-    return payload
-
-
-def workspace_check_result_to_json(
-    result: WorkspaceProjectCheckResult,
-    workspace_manifest: WorkspaceManifest | None,
-) -> dict[str, Any]:
-    item: dict[str, Any] = {
-        "name": result.name,
-        "status": result.status,
-        "path": str(result.root),
-        "manifest_path": str(result.manifest_path) if result.manifest_path is not None else None,
-        "manifest": result.manifest,
-        "checks": [workspace_check_item_to_json(check) for check in result.checks],
-    }
-    if workspace_manifest is not None:
-        item.update(workspace_manifest_item_metadata(result))
-    return item
-
-
-def workspace_manifest_item_metadata(
-    item: WorkspaceProjectStatus | WorkspaceProjectCheckResult,
-) -> dict[str, Any]:
-    metadata: dict[str, Any] = {
-        "repository": item.repository or item.root.name,
-        "expected": item.expected,
-        "required": item.required,
-        "repo": item.repo,
-    }
-    if item.url is not None:
-        metadata["url"] = item.url
-    if item.default_branch is not None:
-        metadata["default_branch"] = item.default_branch
-    return metadata
-
-
-def workspace_project_count(
-    items: tuple[WorkspaceProjectStatus, ...] | tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None,
-) -> int:
-    if workspace_manifest is None:
-        return len(items)
-    return sum(1 for item in items if item.manifest in ("valid", "invalid"))
-
-
-def add_workspace_manifest_json(
-    payload: dict[str, Any],
-    items: tuple[WorkspaceProjectStatus, ...] | tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None,
-) -> None:
-    if workspace_manifest is None:
-        return
-    payload["workspace_manifest"] = {
-        "path": str(workspace_manifest.path),
-        "name": workspace_manifest.name,
-        "schema_version": workspace_manifest.schema_version,
-    }
-    payload["repository_count"] = len(items)
-
-
-def workspace_check_item_to_json(check: ArtifactCheck) -> dict[str, Any]:
-    return check_to_json(check)
-
-
-def last_check_to_json(last_check: ProjectLastCheck | None) -> dict[str, str] | None:
-    if last_check is None:
-        return None
-    return {
-        "checked_at": last_check.checked_at,
-        "status": last_check.status,
-    }
-
-
-def last_check_display(last_check: ProjectLastCheck | None) -> str:
-    if last_check is None:
-        return "-"
-    if len(last_check.checked_at) >= 10:
-        return last_check.checked_at[:10]
-    return last_check.checked_at
-
-
-def print_workspace_status(
-    workspace_root: Path,
-    statuses: tuple[WorkspaceProjectStatus, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> None:
-    if workspace_manifest is not None:
-        print_manifest_workspace_status(workspace_root, statuses, workspace_manifest)
-        return
-
-    print(f"Workspace: {workspace_root} ({len(statuses)} projects)")
-    print()
-    if not statuses:
-        print("No Base-managed projects discovered.")
-        return
-
-    print(f"{'PROJECT':<20} {'STATUS':<6} {'VENV':<8} {'MANIFEST':<8} {'LAST CHECK':<10} PATH")
-    for status in statuses:
-        print(
-            f"{status.name:<20} "
-            f"{status.status:<6} "
-            f"{status.venv:<8} "
-            f"{status.manifest:<8} "
-            f"{last_check_display(status.last_check):<10} "
-            f"{status.root}"
-        )
-
-    attention_count = sum(1 for status in statuses if status.status != "ok")
-    if attention_count:
-        print(f"\n{attention_count} project(s) need attention. Run 'basectl doctor <project>' for details.")
-    else:
-        print("\nAll discovered projects look ok.")
-
-
-def print_manifest_workspace_status(
-    workspace_root: Path,
-    statuses: tuple[WorkspaceProjectStatus, ...],
-    workspace_manifest: WorkspaceManifest,
-) -> None:
-    print(f"Workspace: {workspace_root} ({len(statuses)} repositories)")
-    print(f"Workspace manifest: {workspace_manifest.path} ({workspace_manifest.name})")
-    print()
-    if not statuses:
-        print("No repositories reported by the workspace manifest.")
-        return
-
-    print(
-        f"{'REPOSITORY':<20} {'STATUS':<6} {'REQUIRED':<8} {'REPO':<8} "
-        f"{'VENV':<14} {'MANIFEST':<8} {'LAST CHECK':<10} PATH"
-    )
-    for status in statuses:
-        print(
-            f"{status.repository or status.root.name:<20} "
-            f"{status.status:<6} "
-            f"{yes_no(status.required):<8} "
-            f"{status.repo:<8} "
-            f"{status.venv:<14} "
-            f"{status.manifest:<8} "
-            f"{last_check_display(status.last_check):<10} "
-            f"{status.root}"
-        )
-
-    attention_count = sum(1 for status in statuses if status.status != "ok")
-    if attention_count:
-        print(f"\n{attention_count} repositories need attention. Run 'basectl workspace doctor' for details.")
-    else:
-        print("\nAll workspace repositories look ok.")
-
-
-def print_workspace_check(
-    workspace_root: Path,
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> None:
-    item_name = "repositories" if workspace_manifest is not None else "projects"
-    print(f"Workspace check: {workspace_root} ({len(results)} {item_name})")
-    if workspace_manifest is not None:
-        print(f"Workspace manifest: {workspace_manifest.path} ({workspace_manifest.name})")
-    print_workspace_check_results(results, workspace_manifest)
-
-
-def print_workspace_doctor(
-    workspace_root: Path,
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> None:
-    item_name = "repositories" if workspace_manifest is not None else "projects"
-    print(f"\nWorkspace doctor: {workspace_root} ({len(results)} {item_name})")
-    if workspace_manifest is not None:
-        print(f"Workspace manifest: {workspace_manifest.path} ({workspace_manifest.name})")
-    print_workspace_check_results(results, workspace_manifest)
-
-
-def print_workspace_check_results(
-    results: tuple[WorkspaceProjectCheckResult, ...],
-    workspace_manifest: WorkspaceManifest | None = None,
-) -> None:
-    if not results:
-        if workspace_manifest is None:
-            print("\nNo Base-managed projects discovered.")
-        else:
-            print("\nNo repositories reported by the workspace manifest.")
-        return
-
-    label = "Repository" if workspace_manifest is not None else "Project"
-    for result in results:
-        name = result.repository or result.name
-        print(f"\n{label}: {name} [{result.status}]")
-        print(f"Path: {result.root}")
-        for check in result.checks:
-            print_doctor_finding(doctor_status(check), check.finding_id, check.name, check.message, check.fix)
-
-    error_count = workspace_error_count(results)
-    if error_count:
-        print(f"\nWorkspace has {error_count} error finding(s).")
-        return
-
-    warn_count = sum(1 for result in results for check in result.checks if doctor_status(check) == "warn")
-    if warn_count:
-        print(f"\nWorkspace has {warn_count} warning finding(s).")
-    elif workspace_manifest is not None:
-        print("\nAll workspace repositories passed.")
-    else:
-        print("\nAll discovered projects passed.")
-
-
-def yes_no(value: bool) -> str:
-    return "yes" if value else "no"
-
-
-def dumps_json(payload: dict[str, Any]) -> str:
-    return json.dumps(payload, indent=2)
